Move console argument parsing out of ConsoleModuleLoader
 - Will allow improved testing
 - Temporarily breaks unit tests
