+++ srcStatic/op2ext-Internal.cpp
@@ -1,8 +1,10 @@
 #include "op2ext-Internal.h"
+#include "FileSystemHelper.h"
+#include "ConsoleArgumentParser.h"
 
 bool modulesRunning = false;
 
 Logger logger;
 VolList volList;
-ConsoleModuleLoader consoleModLoader;
+ConsoleModuleLoader consoleModLoader(ParseCommandLine());
 IniModuleLoader iniModuleLoader;
